+++ CHANGELOG.md
@@ -8,6 +8,7 @@ description: |-
 # VKCS Provider's changelog
 ### v0.11.3 (unrealised)
 - Add external_network_id create option for Kubernetes cluster
+- Fix validation for postgrespro_enterprise datastore
 
 #### v0.11.2
 - Support stack_id parameter for vkcs_dataplatform_cluster creation